campus admin automatically generated with school creation

+++ app/models/admin.rb
@@ -1,12 +1,19 @@
 class Admin < ActiveRecord::Base
   belongs_to :user
-  validates :user, presence: true
 
   after_create :change_user_role_to_admin
 
   def change_user_role_to_admin
     self.user.update(role: 'admin')
-    self.user.tutor.destroy
+    if self.user.tutor
+      self.user.tutor.destroy
+    end
+    if self.user.student
+      self.user.student.destroy
+    end
+    if self.user.campus_manager
+      self.user.campus_manager.destroy
+    end
   end
 
 end
